requirements.txt
python-dotenv
python-telegram-bot
httpx
